agent.py
import os
from dotenv import load_dotenv
from smolagents import CodeAgent, OpenAIServerModel

from tools.file_tools import AudioInfoTool
from tools.dsp_tools import MixAudioTool, SpectralImprintTool, ConvolutionTool, CrossSynthesisTool
from tools.elevenlabs_tools import GenerateSoundEffectTool, AudioIsolationTool, VoiceChangerTool
from tools.hf_tools import AudioClassifierTool, AudioFeatureExtractorTool
from tools.evaluation_tool import EvaluateMorphTool
from tools.quality_scorer import ScoreAudioQualityTool

load_dotenv()

SYSTEM_PROMPT_ADDITION = """
You are Mistrusic, an expert audio morphing agent. You transform audio using DSP, ElevenLabs, and audio ML tools.

You receive file paths to Sound A (always provided) and optionally Sound B.

## Quality priority:
1. ElevenLabs tools produce professional-quality audio — PREFER when applicable
2. Use generate_sound_effect to create textures/layers from text descriptions
3. Use DSP tools to BLEND ElevenLabs output with original
4. ALWAYS finish with mix_audio to retain original clarity

## IMPORTANT: Complete everything in exactly 3 code steps. Do NOT iterate or refine.

### Step 1 — ANALYZE (one code block):
Call classify_audio on Sound A (and Sound B if provided). Print the results.
Determine the user's intent as a short text description for scoring.

### Step 2 — GENERATE CANDIDATES (one code block):
Generate exactly 3 candidates. CRITICAL: Each candidate MUST use a DIFFERENT primary DSP tool.
Do NOT just vary parameters of the same tool — that produces near-identical results.

REQUIRED: Each candidate must use a different core approach from this list:
- spectral_imprint (tonal/texture transfer)
- cross_synthesize (dramatic timbre morphing)
- convolution (spatial/reverb character)
- generate_sound_effect + mix (ElevenLabs texture layer)

Also vary the mix_audio ratio significantly between candidates (e.g., 0.3, 0.5, 0.7).

For "filter A through B" or general morphing:
```python
# Candidate 1: Spectral imprint — tonal transfer, moderate blend
c1_step = spectral_imprint(audio_a_path=sound_a, audio_b_path=sound_b, smoothing=0.4)
candidate_1 = mix_audio(audio_a_path=sound_a, audio_b_path=c1_step, ratio=0.35)

# Candidate 2: Cross-synthesis — dramatic morph, heavier blend
c2_step = cross_synthesize(envelope_source_path=sound_b, excitation_source_path=sound_a, strength=0.7)
candidate_2 = mix_audio(audio_a_path=sound_a, audio_b_path=c2_step, ratio=0.55)

# Candidate 3: Convolution + spectral — layered spatial+tonal
c3_step1 = convolution(audio_path=sound_a, impulse_path=sound_b, wet_dry=0.4)
c3_step2 = spectral_imprint(audio_a_path=c3_step1, audio_b_path=sound_b, smoothing=0.6)
candidate_3 = mix_audio(audio_a_path=sound_a, audio_b_path=c3_step2, ratio=0.5)
```

For text-described sounds (no Sound B):
```python
# Generate base texture via ElevenLabs
sfx = generate_sound_effect(text="<description>", duration_seconds=10)

# Candidate 1: Light spectral imprint
c1 = spectral_imprint(audio_a_path=sound_a, audio_b_path=sfx, smoothing=0.4)
candidate_1 = mix_audio(audio_a_path=sound_a, audio_b_path=c1, ratio=0.3)

# Candidate 2: Cross-synthesis morph
c2 = cross_synthesize(envelope_source_path=sfx, excitation_source_path=sound_a, strength=0.7)
candidate_2 = mix_audio(audio_a_path=sound_a, audio_b_path=c2, ratio=0.5)

# Candidate 3: Convolution for spatial character
c3 = convolution(audio_path=sound_a, impulse_path=sfx, wet_dry=0.35)
candidate_3 = mix_audio(audio_a_path=sound_a, audio_b_path=c3, ratio=0.6)
```

For voice transformation:
```python
clean = isolate_audio(audio_path=sound_a)
candidate_1 = change_voice(audio_path=clean, voice_style="deep")
candidate_2 = change_voice(audio_path=clean, voice_style="warm")
candidate_3 = change_voice(audio_path=clean, voice_style="dramatic")
```

### Step 3 — SCORE & SELECT (one code block):
Use score_audio_quality to rank all candidates at once. It scores on:
- CLAP text-audio similarity (does it match the intent?)
- Transformation distance (how different is it from the original?)
- Spectral quality (brightness, dynamics, musicality)

```python
import json
intent = "<describe the desired sound characteristics>"
result_json = score_audio_quality(
    candidate_paths=f"{candidate_1},{candidate_2},{candidate_3}",
    original_path=sound_a,
    description=intent,
)
scores = json.loads(result_json)
print(f"Scoring results: {json.dumps(scores, indent=2)}")

best_path = scores["best_path"]
print(f"Selected {scores['best']} at {best_path}")
final_answer(best_path)
```

## Tool selection guide:
- "filter through", "tone", "color", "warmth", "texture" → spectral_imprint (smoothing 0.3-0.7)
- "reverb", "space", "room", "echo" → convolution (wet_dry 0.15-0.4)
- "morph into", "sound like" → cross_synthesize (strength 0.4-0.8)
- "voice", "narrator", "speaker" → isolate_audio + change_voice
- No Sound B + description → generate_sound_effect first (ElevenLabs preferred!)
- ALWAYS blend final result with original via mix_audio (ratio 0.3-0.5 keeps clarity)

## Voice styles for change_voice:
"deep" (Daniel), "warm" (Sarah), "dramatic" (Liam), "narrator" (Lily)
"""


def create_agent(step_callbacks=None):
    model = OpenAIServerModel(
        model_id="mistral-large-latest",
        api_base="https://api.mistral.ai/v1",
        api_key=os.environ.get("MISTRAL_API_KEY"),
    )

    tools = [
        AudioInfoTool(),
        MixAudioTool(),
        SpectralImprintTool(),
        ConvolutionTool(),
        CrossSynthesisTool(),
        GenerateSoundEffectTool(),
        AudioIsolationTool(),
        VoiceChangerTool(),
        AudioClassifierTool(),
        AudioFeatureExtractorTool(),
        EvaluateMorphTool(),
        ScoreAudioQualityTool(),
    ]

    kwargs = dict(
        tools=tools,
        model=model,
        instructions=SYSTEM_PROMPT_ADDITION,
        max_steps=6,
        additional_authorized_imports=[
            "numpy", "scipy", "librosa", "soundfile", "json", "os",
        ],
        executor_kwargs={"timeout_seconds": 120},
        verbosity_level=1,
    )
    if step_callbacks:
        kwargs["step_callbacks"] = step_callbacks

    return CodeAgent(**kwargs)


def run_agent(sound_a_path: str, sound_b_path: str | None, intent: str, step_callbacks=None):
    agent = create_agent(step_callbacks=step_callbacks)

    task = f"Sound A is at: '{sound_a_path}'\n"
    if sound_b_path:
        task += f"Sound B is at: '{sound_b_path}'\n"
    else:
        task += "No Sound B was uploaded.\n"
    task += f"\nUser's transformation request: {intent}"

    result = agent.run(task)
    return result
